Plan 044 Step 4: split types.ts behind seam-aligned barrel
Move types into src/types/* by seam; types.ts re-exports all so from './types'
import sites stay unchanged.
